meraki/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
exceptiongroup==1.2.2
Flask==3.1.0
frozenlist==1.5.0
future==1.0.0
idna==3.10
iniconfig==2.1.0
ipcalc==1.99.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
meraki==1.56.0
multidict==6.2.0
netmiko==4.5.0
ntc_templates==7.8.0
packaging==24.2
paramiko==3.5.1
pluggy==1.5.0
propcache==0.3.1
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
pyproject_hooks==1.2.0
pyserial==3.5
pytest==7.4.4
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
resolvelib==1.0.1
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scp==0.15.0
six==1.17.0
textfsm==1.1.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
